backend/fashion_detection_training/requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
Pillow>=9.0.0
PyYAML>=6.0
tqdm>=4.65.0
